Ensure tx submission errors are actually surfaced
When the TransactionSubmitter returned an error during a compliance-based payment, the error would get ignored because the `err` variable in the main scope was shadowed by another in the scope where the submission was made. This meant that the HTTP request to the payment endpoint would return a 200 success (just with bad data `{ledger: null}`) and, additionally, it prevented the error from ever being logged, so the server administrator can't see what went wrong, either.

The lack of logs is actually something that completely stopped my progress trying to use the bridge server for about a full day, scratching my head trying to understand what was going wrong.